Make MentoringRequest a common type
