+++ app/javascript/components/mentoring/request/StartMentoringPanel.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useMutation } from 'react-query'
 import { sendRequest } from '../../../utils/send-request'
 import { typecheck } from '../../../utils/typecheck'
-import { MentoringRequest } from '../Session'
+import { MentoringRequest } from '../../types'
 import { useIsMounted } from 'use-is-mounted'
 import { Loading } from '../../common'
 import { ErrorBoundary, useErrorHandler } from '../../ErrorBoundary'
